refactor: modularize controller utilities
