+++ Controller/config.py
@@ -1,0 +1,7 @@
+GPU_MEMORY_LABEL = "nvidia.com/gpu.memory"
+IN_CLUSTER = True
+SERVICE_FILE = './information/service.json'
+SERVICESPEC_FILE = './information/serviceSpec.json'
+SUBSCRIPTION_FILE = './information/subscription.json'
+NODE_STATUS_FILE = './information/nodestatus.json'
+LOG_FILE = './logdir/controller.log'
